test-requirements.txt
mock
pytz
python-dateutil==2.4.2
coverage
coveralls
Fabric==1.10.2
pycrypto==2.6.1
celery==3.1.19
django-celery==3.1.17
